chore: move mocks to internal
