+++ packages/shared/common/src/index.ts
@@ -10,6 +10,5 @@ export * from './utils';
 
 export * as internal from './internal';
 export * from './errors';
-export * as mocks from './mocks';
 
 export { AttributeReference, Context, ContextFilter };
