+++ packages/shared/common/src/internal/index.ts
@@ -1,1 +1,2 @@
 export * from './events';
+export * as mocks from './mocks';
