+++ packages/shared/common/src/internal/mocks/clientContext.ts
@@ -1,5 +1,5 @@
-import { createSafeLogger } from '../logging';
-import { ClientContext } from '../options';
+import { createSafeLogger } from '../../logging';
+import { ClientContext } from '../../options';
 import basicPlatform from './platform';
 
 const clientContext = new ClientContext(
