+++ packages/shared/sdk-server/__tests__/LDClient.allFlags.test.ts
@@ -1,9 +1,11 @@
-import { mocks } from '@launchdarkly/js-sdk-common';
+import { internal } from '@launchdarkly/js-sdk-common';
 
 import { LDClientImpl } from '../src';
 import TestData from '../src/integrations/test_data/TestData';
 import TestLogger, { LogLevel } from './Logger';
 import makeCallbacks from './makeCallbacks';
+
+const { mocks } = internal;
 
 const defaultUser = { key: 'user' };
