+++ packages/shared/sdk-server/__tests__/LDClient.events.test.ts
@@ -1,8 +1,10 @@
-import { Context, internal, mocks } from '@launchdarkly/js-sdk-common';
+import { Context, internal } from '@launchdarkly/js-sdk-common';
 
 import { LDClientImpl } from '../src';
 import TestData from '../src/integrations/test_data/TestData';
 import makeCallbacks from './makeCallbacks';
+
+const { mocks } = internal;
 
 const defaultUser = { key: 'user' };
 const anonymousUser = { key: 'anon-user', anonymous: true };
