+++ packages/shared/sdk-server/__tests__/LDClientImpl.bigSegments.test.ts
@@ -1,4 +1,4 @@
-import { Crypto, Hasher, Hmac, mocks } from '@launchdarkly/js-sdk-common';
+import { Crypto, Hasher, Hmac, internal } from '@launchdarkly/js-sdk-common';
 
 import { BigSegmentStore } from '../src/api/interfaces';
 import { LDBigSegmentsOptions } from '../src/api/options/LDBigSegmentsOptions';
@@ -7,6 +7,8 @@ import TestData from '../src/integrations/test_data/TestData';
 import LDClientImpl from '../src/LDClientImpl';
 import { makeSegmentMatchClause } from './evaluation/flags';
 import makeCallbacks from './makeCallbacks';
+
+const { mocks } = internal;
 
 const user = { key: 'userkey' };
 const bigSegment = {
