+++ packages/shared/sdk-server/__tests__/evaluation/Evaluator.bucketing.test.ts
@@ -1,9 +1,11 @@
-import { Context, mocks } from '@launchdarkly/js-sdk-common';
+import { Context, internal } from '@launchdarkly/js-sdk-common';
 
 import { Flag } from '../../src/evaluation/data/Flag';
 import { Rollout } from '../../src/evaluation/data/Rollout';
 import Evaluator from '../../src/evaluation/Evaluator';
 import noQueries from './mocks/noQueries';
+
+const { mocks } = internal;
 
 const evaluator = new Evaluator(mocks.basicPlatform, noQueries);
