+++ packages/shared/sdk-server/__tests__/evaluation/Evaluator.clause.test.ts
@@ -1,4 +1,4 @@
-import { AttributeReference, Context, LDContext, mocks } from '@launchdarkly/js-sdk-common';
+import { AttributeReference, Context, internal, LDContext } from '@launchdarkly/js-sdk-common';
 
 import { Clause } from '../../src/evaluation/data/Clause';
 import { Flag } from '../../src/evaluation/data/Flag';
@@ -10,6 +10,8 @@ import {
   makeClauseThatMatchesUser,
 } from './flags';
 import noQueries from './mocks/noQueries';
+
+const { mocks } = internal;
 
 const evaluator = new Evaluator(mocks.basicPlatform, noQueries);
