+++ packages/shared/sdk-server/__tests__/evaluation/Evaluator.rules.test.ts
@@ -1,6 +1,6 @@
 // Tests of flag evaluation at the rule level. Clause-level behavior is covered
 // in detail in Evaluator.clause.tests and (TODO: File for segments).
-import { AttributeReference, Context, LDContext, mocks } from '@launchdarkly/js-sdk-common';
+import { AttributeReference, Context, internal, LDContext } from '@launchdarkly/js-sdk-common';
 
 import { Clause } from '../../src/evaluation/data/Clause';
 import { Flag } from '../../src/evaluation/data/Flag';
@@ -12,6 +12,8 @@ import {
   makeFlagWithRules,
 } from './flags';
 import noQueries from './mocks/noQueries';
+
+const { mocks } = internal;
 
 const basicUser: LDContext = { key: 'userkey' };
 const basicSingleKindUser: LDContext = { kind: 'user', key: 'userkey' };
