+++ packages/shared/sdk-server/__tests__/evaluation/Evaluator.test.ts
@@ -1,10 +1,12 @@
-import { Context, LDContext, mocks } from '@launchdarkly/js-sdk-common';
+import { Context, internal, LDContext } from '@launchdarkly/js-sdk-common';
 
 import { Flag } from '../../src/evaluation/data/Flag';
 import EvalResult from '../../src/evaluation/EvalResult';
 import Evaluator from '../../src/evaluation/Evaluator';
 import Reasons from '../../src/evaluation/Reasons';
 import noQueries from './mocks/noQueries';
+
+const { mocks } = internal;
 
 const offBaseFlag = {
   key: 'feature0',
